Editable automations section on Edit policy page
- Extract the automations table + Learn more + Continuous-retry checkbox
  into a shared PolicyAutomationsFields component (in
  pages/policies/components/PolicyAutomationsFields/), used by both the
  ManageAutomationsModal and the Edit policy page.
- Render the editable automations section in PolicyForm between Target and
  Critical, gated on automationsConfig being loaded so the initial state
  (incl. the webhook/ticket row) initializes correctly on a cold refresh.
- Wire automations into the page's existing Save: per-policy fields +
  webhook/ticket config persist via useUpdatePolicyAutomations alongside
  the core policy update.
- Remove the Cancel button (the back button already does the same) and
  rename "Run" -> "Run policy".
- Unify getTicketOrWebhookState + getWebhookOrTicketPolicyIds into a
  single getTicketOrWebhookInfo helper returning { state, policyIds }.
- Make getAutomationsPayload pure (returns error string instead of
  flashing), so the shared component no longer depends on
  NotificationContext.
