+++ frontend/pages/policies/components/PolicyAutomationsFields/index.ts
@@ -1,0 +1,13 @@
+export { default } from "./PolicyAutomationsFields";
+export type {
+  IPolicyAutomationsFieldsHandle,
+  IPolicyAutomationsPayload,
+} from "./PolicyAutomationsFields";
+
+export { default as useUpdatePolicyAutomations } from "./hooks/useUpdatePolicyAutomations";
+export type {
+  IPolicyAutomationUpdate,
+  IUpdatePolicyAutomationsVars,
+} from "./hooks/useUpdatePolicyAutomations";
+
+export { getTicketOrWebhookInfo } from "./helpers";
